server/backend/requirements.txt
fastapi
uvicorn[standard]
psycopg[binary]
python-dotenv
pytest
requests
